Restructured Application Generation
- Created factory method to be used with WGSI servers
- Corrected imports
- Restructured file heirarchy
