+++ app/core/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+core = Blueprint('core', __name__)
+
+from . import routes
+from ..dependencies import sockets
