+++ app/static/icons/palette.png
