+++ app/static/js/functions.js
