+++ app/static/js/nav.js
