+++ app/templates/manual.html
