+++ app/templates/monitor.html
